NOVA-1171 refactor project interfaces
Introduce a common project for common models.

Sever the dependency between Data and Client projects.

Stop using a client API model in the inner workings of the matching,
i.e. replace uses of PotentialMatch with the new PotentialSearchResult.
This is so that filtering and scoring data (such as hla) can be
included in matching results, without sullying the client API.