NOVA-NONE: Load all donors into memory, rather than keeping open EF connection to SQL
We're hoping this will prevent the gradual slowdown of the hla update job we're seeing on dev.
Additionally it seems like a bad idea to have a single connection open via the EF queriable for the runtime of this job (multiple hours).

If memory proves to be an issue we should implement efficient batching to fetch all donors