NOVA-1289: Merge matcing test fixtures for single loci into single, parameterised test fixture
On balance, I decided that the added complexity wasn't huge, could be encapsulated in the builder classes,
and the removal of a large amount of duplication of tests cases across different loci was more important.

If loci behaviuor differs in future, we can add separate fixtures for them, and remove the parameter for that locus from this file (e.g. DPB1, which isn't present here as it has different behaviuor)