NOVA-NONE: Reduce test data file names
Depending on the path of the solution, the longest of these files were becoming so long that the full path of the files was too long for the build.

This is the quickest fix available, and seems sensible as the prefixes in the file names are also present in the folder names.