NOVA-1444: Move shared WmdaRepository to shared cache, set up within NUnit lifecycle
This had the same issues as the preCalculationService, in that it was set up vefore all tests, regardless of whether they used it.
The effect wasn't as pronounced, but still wortgh fixing.

This has a bonus effect of making the test suite understood sensibly by resharper, so all tests will now run and display results in a well structured format :D