NOVA-1459: Use long for auto-increment id, not int
Currently we're at about 200 million records, just under a tenth of the max int size. As
(a) New ids are generated when we update match data, as we drop and re-create for performance reasons
(b) we want to eventually support a dataset 25x as big

We believe we're going to run out of ints.
As we don't use th id itself, another alternative is to remove the id column and add a composite PK. We don't know if this would have an impact on performance, and for now have chosen the simpler solution.