NOVA-1567: Add p-group match level for four-field alleles
This method currently has two flaws:

(a) While we're specifying the match level as p-group, all this guarantees is that the donor/patient alleles are not identical. It's very possible the match *grade* will be e.g. cDNA, not P-Group, as might be implied by the name
(b) The donor + patient hla selection currently happens on a per position basis, meaning that the selected hla are only guaranteed not to be an allele level direct match - they could still be an allele level cross match...