NOVA-1703: Ensure three field (not fourth field) matches cannot select data with a cross 4-field match
e.g. if a group of valid alleles consists of two values, 15:01:01:01, and 15:01:01:02
It was previoulsy possible for the donor to be typed at a locus: (15:01:01:01, 15:01:01:02), leaving the patient with no valid alleles to choose that will not have a 4-field match

We cannot solve this by requiring groups of at least 3 alleles sharing the same first 3 fields, as we do not have enough TGS test data at some loci.

Instead the fix is to restrict the available donor alleles in this circumstance