NOVA-1818: Make null allele in string flag static
Required that the base converter has a parameterless constructor for unit tests, so we cannot inject this flag.
This means we cannot test the scenario with the flag off, which is OK for now as more development work would be necessary in matching to allow consideration of null alleles within allele strings.