NOVA-1748: Remove old method of adding static test cases
Now that there is a public api for adding donors, we do not need to add all donors for etsting up front, and it becomes much simpler to define donor HLA at the time of writing a test.
The one existing specific test case was no longer in use, so can stay removed.

If we need to re-add this test case in future, we should use the new, simpler static test data provider, and define the specific test cases' data in the test case