fix: ATLAS-366: Squash donor import SQL migrations
See https://github.com/dotnet/efcore/issues/12911 for known issue with renaming indexes.
"Execute" Workaround had caused a state where migrations did not work on a fresh database.

Rather than pouring more effort into the workaround, we are squashing all migrations into one, as at this point in time the system is not live.

Deployed test/dev environments can be fixed by either dropping and recreating the DB, or by replacing the contents of the migrations table with the new migration id (the snapshot of the DB is the same, so this is safe to do at this point in time.)