refactor: ATLAS-590: Use WhenEach instead of WhenAll when converting Haplotype frequencies
Purpose of this is two-fold:

(a) It prevents us opening too many connections when in a deployed Azure environment against an unwarmed HMD cache (technically the caching extensions used in HMD will also protect against this)

(b) It allows us to fail early when a set is in a wrong format. e.g. when I uploaded a csv with the DRB and DQB headers swapped, this import process hung for a very long time as it processed ~20,000 HMD lookup errors. This way we will throw an exception after the first failed lookup.