fix: ATLAS-665: Remove unnecessary dependency of Functions on Functions.PUiblicApi
Due to the nature of [assembly:], this includsion meant that the `Startup` file for `Atlas.Functions.PublicApi` was being run erroneously - and overriding the settings we wanted in Atlas.Functions!

This comment stands as a warning: *do not let functions app depend on each other*

It also adds another justification for the new, independent models package.

qq service bus