Convert HF sets from CSV to Json
Run the Atlas.MatchPrediction.Functions app and hit the 'ConvertToJson' endpoint. There are two keys:

filePath : Full path to the csv that you want to convert to a json [REQUIRED]

outputPath : Full path to where you want to save the new json file to. If left back the full json will just be returned [OPTIONAL]