fix: ATLAS-699: Remove cleanup notification for data refresh
Since moving to an elastic plan, this notification is no longer useful.
It triggers when scaling out to a new instance, which happens very soon after the refresh starts due to high CPU usage.

The "RunOnStartup" is not recommended for live applications by Microsoft.

We could get this to keep working by migrating just the data refresh to an app service plan, but provisioning a new app service plan just for this will be more detrimental than removing the notification, as:
- It will mean paying for an app service that isn't being used for most of the year
- Being on a scalable plan should hopefully solve some of the reasons the job has been interrupted in development anyway (due to service worker maintenance in Azure).
- Start/End notifications, combined with AI logging, should give anyone supporting the system enough information about when data refreshes stall.