performance: #682: Calculate likelihoods for diplotypes in series
Task.WhenAll might hinder more than it helps here: we're likely to have a *large* amount of tasks to manage (at least 1000s).
When running in the cloud we may not be able to make the most of parallel threading due to the ability to run multiple functions on the server at once.
Additionally, the asynchronous part of this call is a cached database call - meaning we're not going to spend a lot of time waiting for external resources here (and even if we were, our database might not be happy with thousands of concurrent connections!)